manage ipcp register in userspace

+++ tests/epilogue.sh
@@ -1,1 +1,3 @@
-kill -SIGINT $(pgrep ipcm)
+if [ "$batch" != "0" ]; then
+    kill -SIGINT $(pgrep ipcm)
+fi
